modules/visual_model.py
import numpy as np
import os
import asyncio

try:
    import faiss
    FAISS_AVAILABLE = True
except ImportError:
    FAISS_AVAILABLE = False
    print("Warning: FAISS not available. Visual similarity check will be disabled.")

try:
    import onnxruntime as ort
    ONNX_AVAILABLE = True
except ImportError:
    ONNX_AVAILABLE = False

class VisualUrlDetector:
    """
    Visual similarity detector using Siamese Networks and FAISS.
    """
    def __init__(self, index_path=None, model_path=None):
        self.index = None
        self.model_session = None
        self.index_path = index_path or os.path.join('data', 'trusted_brands.index')
        self.model_path = model_path or os.path.join('data', 'visual_model.onnx')
        
        self._initialize()

    def _initialize(self):
        if FAISS_AVAILABLE and os.path.exists(self.index_path):
            try:
                self.index = faiss.read_index(self.index_path)
                print(f"Loaded visual index with {self.index.ntotal} vectors.")
            except Exception as e:
                print(f"Error loading FAISS index: {e}")
        
        if ONNX_AVAILABLE and os.path.exists(self.model_path):
            try:
                self.model_session = ort.InferenceSession(self.model_path)
            except Exception as e:
                print(f"Error loading Visual ONNX model: {e}")

    async def compute_embedding(self, image_data):
        """
        Compute vector embedding for a screenshot (Simulated for zero-dependency Phase 1 environment).
        Real implementation would preprocess image -> ONNX -> Vector.
        """
        if not self.model_session:
            # Fallback/Placeholder: Return random vector
            return np.random.rand(1, 128).astype('float32')
        
        # Real logic (commented out until image libraries verified)
        # img = preprocess(image_data)
        # embedding = self.model_session.run(None, {'input': img})[0]
        # return embedding
        
        return np.random.rand(1, 128).astype('float32')

    async def find_similarity(self, image_path):
        """
        Find visually similar brands in the database.
        """
        if not self.index or not FAISS_AVAILABLE:
            return None
        
        # 1. Get embedding
        embedding = await self.compute_embedding(image_path)
        
        # 2. Search FAISS
        # D = distances, I = indices
        D, I = self.index.search(embedding, 1) 
        
        closest_dist = D[0][0]
        closest_idx = I[0][0]
        
        # Threshold (arbitrary for demo)
        if closest_dist < 0.5:
            # Map index to brand name (would need a JSON map)
            return "Unknown Brand (Visual Match)"
            
        return None

# Global instance
visual_detector = VisualUrlDetector()
