requirements.txt
# ==============================================================================
# URL Sentinel: Hypersonic Zero (Titan Edition)
# The <1ms Anti-Phishing System
# ==============================================================================

# --- Core Async Framework (Phase 1) ---
quart==0.19.4
quart-cors==0.7.0
uvloop==0.19.0; sys_platform != 'win32' # Accelerated event loop (Linux/Mac)

# --- Legacy/Fallback ---
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
dnspython==2.4.2

# --- AI & Vision (Phase 2 & Titan) ---
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.2.0
joblib==1.3.2
onnxruntime==1.17.0
pillow==10.2.0
pytesseract==0.3.10
faiss-cpu==1.7.4

# --- Rust Integration (Phase 1++) ---
maturin==1.4.0
# Note: 'rust_core' is built locally via maturin

# --- Testing ---
pytest==7.4.3
pytest-asyncio==0.23.5